Refactor: Move LegacySettingsReader into the data package
SettingsRepository (data) reached up into viewmodel to read the legacy
individual-key settings. The reader only touches SharedPreferences and
AppSettings, so it belongs beside its one caller; the move also drops
twelve now-redundant imports across the reader and its test.
